server/pyproject.toml
[project]
name = "python-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich[toml]>=0.9.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "polyfactory>=2.21.0",
    "pydantic>=2.11.5",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.34.3",
]

[tool.ruff]
indent-width = 2
line-length = 80

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
  "I",  # isort (import sorting)
  "C90", # mccabe complexity
  "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
lines-between-types = 0

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
